Refactored and all tests working.
