+++ runtime/entity/entity.go
@@ -1,4 +1,4 @@
-package operationsbus
+package entity
 
 // This is the entity that will be grabbed from the database by each implementer. They will have to use this GetLatestOperationId() method when guarding for concurrency.
 type Entity interface {
