+++ runtime/matcher/matcher_test.go
@@ -1,21 +1,24 @@
-package operationsbus
+package matcher
 
 import (
 	"context"
 	"fmt"
 	"reflect"
 	"testing"
+
+	"github.com/Azure/aks-async/runtime/entity"
+	"github.com/Azure/aks-async/runtime/operation"
 )
 
 func TestMatcher(t *testing.T) {
 	matcher := NewMatcher()
 
-	operation := "LongRunning"
-	matcher.Register(operation, &LongRunning{})
+	operationName := "LongRunning"
+	matcher.Register(operationName, &LongRunning{})
 
-	retrieved, exists := matcher.Get(operation)
+	retrieved, exists := matcher.Get(operationName)
 	if !exists {
-		t.Fatalf("Operation %s should exist in the matcher, instead got: %t", operation, exists)
+		t.Fatalf("Operation %s should exist in the matcher, instead got: %t", operationName, exists)
 	}
 
 	longRunningOp := &LongRunning{}
@@ -25,7 +28,7 @@ func TestMatcher(t *testing.T) {
 	}
 
 	// Retrieve an instance of the type associated with the key operation
-	instance, err := matcher.CreateOperationInstance(operation)
+	instance, err := matcher.CreateOperationInstance(operationName)
 	if err != nil {
 		fmt.Println("Type not found")
 		return
@@ -37,7 +40,7 @@ func TestMatcher(t *testing.T) {
 	}
 
 	ctx := context.Background()
-	_, _ = instance.InitOperation(ctx, OperationRequest{})
+	_, _ = instance.InitOperation(ctx, operation.OperationRequest{})
 	_ = instance.Run(ctx)
 	if longOp, ok := instance.(*LongRunning); ok {
 		if longOp.num != 2 {
@@ -53,7 +56,7 @@ func TestMatcher_RegisterAndGetEntity(t *testing.T) {
 
 	entityKey := "TestEntity"
 	lastOperationId := "1"
-	matcher.RegisterEntity(entityKey, func(latestOperationId string) Entity {
+	matcher.RegisterEntity(entityKey, func(latestOperationId string) entity.Entity {
 		return &TestEntity{latestOperationId: latestOperationId}
 	})
 
@@ -65,15 +68,15 @@ func TestMatcher_RegisterAndGetEntity(t *testing.T) {
 	// Create an instance of the entity using the registered creator
 	entityInstance := matcher.EntityCreators[entityKey]
 
-	var entity Entity
+	var e entity.Entity
 	if f, ok := matcher.EntityCreators[entityKey]; ok {
-		entity = f(lastOperationId)
+		e = f(lastOperationId)
 	} else {
 		t.Fatalf("Expected entity instance of type *TestEntity. Instead got: %T", entityInstance)
 	}
 
-	if entity.(*TestEntity).latestOperationId != "1" {
-		t.Fatalf("Expected entity name to be %s. Instead got: %s", lastOperationId, entity.(*TestEntity).latestOperationId)
+	if e.(*TestEntity).latestOperationId != "1" {
+		t.Fatalf("Expected entity name to be %s. Instead got: %s", lastOperationId, e.(*TestEntity).latestOperationId)
 	}
 }
 
@@ -82,7 +85,7 @@ func TestMatcher_CreateEntityInstance(t *testing.T) {
 
 	entityKey := "TestEntity"
 	lastOperationId := "1"
-	matcher.RegisterEntity(entityKey, func(latestOperationId string) Entity {
+	matcher.RegisterEntity(entityKey, func(latestOperationId string) entity.Entity {
 		return &TestEntity{latestOperationId: latestOperationId}
 	})
 
@@ -91,10 +94,10 @@ func TestMatcher_CreateEntityInstance(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Expected no error. Instead got: %v", err)
 	}
-	if entity, ok := entityInstance.(*TestEntity); !ok {
+	if testEntityInstance, ok := entityInstance.(*TestEntity); !ok {
 		t.Fatalf("Expected entity instance of type *TestEntity. Instead got: %T", entityInstance)
 	} else {
-		if entity.latestOperationId != lastOperationId {
+		if testEntityInstance.latestOperationId != lastOperationId {
 			t.Fatalf("lastestOperationId of entity doesn't match what was used to create the instance: " + lastOperationId)
 		}
 	}
@@ -137,17 +140,17 @@ type LongRunning struct {
 	num int
 }
 
-var _ ApiOperation = (*LongRunning)(nil)
+var _ operation.ApiOperation = (*LongRunning)(nil)
 
-func (lr *LongRunning) InitOperation(ctx context.Context, req OperationRequest) (ApiOperation, error) {
+func (lr *LongRunning) InitOperation(ctx context.Context, req operation.OperationRequest) (operation.ApiOperation, error) {
 	fmt.Println("Initializing LongRunning operation with request")
 	lr.num = 1
 	return nil, nil
 }
 
-func (lr *LongRunning) GuardConcurrency(ctx context.Context, entity Entity) *CategorizedError {
+func (lr *LongRunning) GuardConcurrency(ctx context.Context, entityInstance entity.Entity) *entity.CategorizedError {
 	fmt.Println("Guarding concurrency in LongRunning operation")
-	return &CategorizedError{}
+	return nil
 }
 
 func (lr *LongRunning) Run(ctx context.Context) error {
@@ -156,6 +159,6 @@ func (lr *LongRunning) Run(ctx context.Context) error {
 	return nil
 }
 
-func (lr *LongRunning) GetOperationRequest() *OperationRequest {
-	return &OperationRequest{}
+func (lr *LongRunning) GetOperationRequest() *operation.OperationRequest {
+	return &operation.OperationRequest{}
 }
