+++ runtime/processor/processor.go
@@ -1,4 +1,4 @@
-package operationsbus
+package processor
 
 import (
 	"errors"
@@ -8,18 +8,23 @@ import (
 
 	oc "github.com/Azure/OperationContainer/api/v1"
 	"github.com/Azure/go-shuttle/v2"
+
+	ec "github.com/Azure/aks-async/runtime/entity_controller"
+	"github.com/Azure/aks-async/runtime/handlers"
+	"github.com/Azure/aks-async/runtime/hooks"
+	"github.com/Azure/aks-async/runtime/matcher"
 )
 
 // The processor will be used to process all the operations using the default values or with handlers set by the user.
 func CreateProcessor(
 	serviceBusReceiver sb.ReceiverInterface,
-	matcher *Matcher,
+	matcher *matcher.Matcher,
 	operationContainer oc.OperationContainerClient,
-	entityController EntityController,
+	entityController ec.EntityController,
 	logger *slog.Logger,
 	customHandler shuttle.HandlerFunc,
 	processorOptions *shuttle.ProcessorOptions,
-	hooks []BaseOperationHooksInterface,
+	hooks []hooks.BaseOperationHooksInterface,
 ) (*shuttle.Processor, error) {
 
 	if serviceBusReceiver == nil {
@@ -33,7 +38,7 @@ func CreateProcessor(
 	// Define the default handler chain
 	// Use the default handler if a custom handler is not provided
 	if customHandler == nil {
-		customHandler = DefaultHandlers(serviceBusReceiver, matcher, operationContainer, entityController, logger, hooks)
+		customHandler = handlers.DefaultHandlers(serviceBusReceiver, matcher, operationContainer, entityController, logger, hooks)
 	}
 
 	// Set default processor options.
